Refacroting solution, moved unit storage into Tile class
