+++ MapEditor/Commands/CommandType.cs
@@ -1,0 +1,14 @@
+﻿namespace MapEditor.Commands
+{
+    public enum CommandType
+    {
+        None = 0,
+        Move,
+        Stop,
+        Undo,
+        Redo,
+        PlaceTile,
+        AddUnit,
+        MoveCamera
+    }
+}
