+++ MapEditor/Commands/ICommand.cs
@@ -1,0 +1,7 @@
+﻿namespace MapEditor.Commands
+{
+    public interface ICommand
+    {
+        CommandType Id { get; }
+    }
+}
