+++ MapEditor/Commands/MoveCommand.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using MapEditor.Engine;
+﻿using MapEditor.Engine;
 
 namespace MapEditor.Commands
 {
@@ -13,10 +8,5 @@ namespace MapEditor.Commands
         public int X { get; set; }
         public int Y { get; set; }
         public bool IsQueued { get; set; }
-    }
-
-    public class StopCommand : ICommand
-    {
-        public CommandType Id { get; } = CommandType.Stop;
     }
 }
