+++ MapEditor/Commands/RedoCommand.cs
@@ -1,0 +1,7 @@
+﻿namespace MapEditor.Commands
+{
+    public class RedoCommand : ICommand
+    {
+        public CommandType Id { get; } = CommandType.Redo;
+    }
+}
