+++ MapEditor/Commands/StopCommand.cs
@@ -1,0 +1,9 @@
+﻿using MapEditor.Engine;
+
+namespace MapEditor.Commands
+{
+    public class StopCommand : ICommand
+    {
+        public CommandType Id { get; } = CommandType.Stop;
+    }
+}
