+++ MapEditor/Commands/UndoCommand.cs
@@ -1,0 +1,7 @@
+﻿namespace MapEditor.Commands
+{
+    public class UndoCommand : ICommand
+    {
+        public CommandType Id { get; } = CommandType.Undo;
+    }
+}
