+++ MapEditor/Commands/UnitComponent.cs
@@ -1,0 +1,10 @@
+﻿using MapEditor.Components;
+
+namespace MapEditor.Commands
+{
+    public class UnitComponent : IComponent
+    {
+        public string Title { get; set; }
+        public string Name { get; set; }
+    }
+}
