+++ MapEditor/Common/Matrix.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MapEditor.Common
+﻿namespace MapEditor.Common
 {
     public class Matrix
     {
