+++ MapEditor/Components/AnimationComponent.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MapEditor.Components
+{
+    //public class AnimationComponent : IComponent
+    //{
+    //    public ComponentType Type { get; } = ComponentType.RenderAnimation;
+    //    public Animation Animation { get; set; }
+    //}
+}
