+++ MapEditor/Components/CollisionComponent.cs
@@ -1,0 +1,10 @@
+﻿using MapEditor.Controllers.CollisionHandler;
+using MapEditor.Entities;
+
+namespace MapEditor.Components
+{
+    public class CollisionComponent : IComponent
+    {
+        public ICollider Collider { get; set; }
+    }
+}
