+++ MapEditor/Components/IComponent.cs
@@ -1,0 +1,6 @@
+﻿namespace MapEditor.Components
+{
+    public interface IComponent
+    {
+    }
+}
