+++ MapEditor/Components/PathingComponent.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using MapEditor.Engine;
+
+namespace MapEditor.Components
+{
+    public class PathingComponent : IComponent
+    {
+        public List<ITarget> Destinations = new List<ITarget>();
+    }
+}
