+++ MapEditor/Components/PhysicsComponent.cs
@@ -1,0 +1,7 @@
+﻿namespace MapEditor.Components
+{
+    public class PhysicsComponent : IComponent
+    {
+        public float Mass { get; set; }
+    }
+}
