+++ MapEditor/Engine/IHandleCommand.cs
@@ -1,0 +1,10 @@
+﻿using MapEditor.Commands;
+
+namespace MapEditor.Engine
+{
+    public interface IHandleCommand
+    {
+        void Handle(ICommand command);
+        void Undo(ICommand command);
+    }
+}
