+++ MapEditor/Engine/TerrainType.cs
@@ -1,0 +1,10 @@
+﻿namespace MapEditor.Engine
+{
+    public enum TerrainType
+    {
+        Empty = 0,
+        Water,
+        Land,
+        ImpassableLand,
+    }
+}
