+++ MapEditor/Entities/Entity.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Newtonsoft.Json;
+using MapEditor.Components;
 
 namespace MapEditor.Entities
 {
