+++ MapEditor/Handlers/CollisionHandler/BoundingCircle.cs
@@ -1,0 +1,22 @@
+﻿using System.Drawing;
+using MapEditor.Common;
+
+namespace MapEditor.Controllers.CollisionHandler
+{
+    public class BoundingCircle : ICollider
+    {
+        public float Radius { get; set; }
+        public Point Position { get; set; }
+
+        public bool IsCollided(ICollider collider)
+        {
+            if (collider is BoundingCircle circle)
+                return this.Contains(circle);
+
+            if (collider is BoundingBox box)
+                return this.Contains(box);
+
+            return false;
+        }
+    }
+}
