+++ MapEditor/Handlers/CollisionHandler/ICollider.cs
@@ -1,0 +1,10 @@
+﻿using System.Drawing;
+
+namespace MapEditor.Controllers.CollisionHandler
+{
+    public interface ICollider
+    {
+        Point Position { get; set; }
+        bool IsCollided(ICollider collider);
+    }
+}
