+++ MapEditor/Handlers/MovementHandler/MovementMode.cs
@@ -1,0 +1,13 @@
+﻿namespace MapEditor.Controllers.MovementHandler
+{
+    // https://gamedevelopment.tutsplus.com/series/understanding-steering-behaviors--gamedev-12732
+    public enum MovementMode
+    {
+        Move = 0,
+        Follow,
+        Intercept,
+        Evade,
+        Roam,
+        Patrol,
+    }
+}
